Connect `iced_web` with `iced` properly
