+++ src/web.rs
@@ -1,0 +1,1 @@
+pub use iced_web::*;
