+++ web/Cargo.toml
@@ -17,8 +17,7 @@ maintenance = { status = "actively-developed" }
 [dependencies]
 iced_core = { version = "0.1.0-alpha", path = "../core" }
 dodrio = "0.1.0"
-futures-preview = "=0.3.0-alpha.18"
-wasm-bindgen = "0.2.50"
+wasm-bindgen = "0.2.51"
 
 [dependencies.web-sys]
 version = "0.3.27"
